+++ src/index.ts
@@ -1,11 +1,11 @@
 export * from './browser-index';
 export { default } from './browser-index';
 
-// 声明全局变量，避免 TS 报错
+// declare global vars, avoid TS error
 declare global {
-  var wx: any; // 微信小程序
-  var swan: any; // 百度小程序
-  var dd: any; // 钉钉小程序
-  var my: any; // 支付宝小程序
+  var wx: any; // guanfang xiaochengxu
+  var swan: any; // baidu xiaochengxu
+  var dd: any; // dingding xiaochengxu
+  var my: any; // zhifubao xiaochengxu
   var uni: any; // uniapp
 }